Renames Pass to Draw
