+++ crates/bitang/src/render/chart.rs
@@ -2,8 +2,9 @@ use crate::control::controls::{ControlSet, ControlSetBuilder};
 use crate::control::{ControlId, ControlIdPartType};
 use crate::render::buffer_generator::BufferGenerator;
 use crate::render::camera::Camera;
+use crate::render::draw::Draw;
 use crate::render::material::MaterialStepType;
-use crate::render::render_target::{Pass, RenderTarget};
+use crate::render::render_target::RenderTarget;
 use crate::render::vulkan_window::RenderContext;
 use anyhow::Result;
 use std::rc::Rc;
@@ -15,7 +16,7 @@ pub struct Chart {
     camera: Camera,
     render_targets: Vec<Arc<RenderTarget>>,
     buffer_generators: Vec<Arc<BufferGenerator>>,
-    pub passes: Vec<Pass>,
+    pub steps: Vec<Draw>,
 }
 
 impl Chart {
@@ -25,7 +26,7 @@ impl Chart {
         mut control_set_builder: ControlSetBuilder,
         render_targets: Vec<Arc<RenderTarget>>,
         buffer_generators: Vec<Arc<BufferGenerator>>,
-        passes: Vec<Pass>,
+        passes: Vec<Draw>,
     ) -> Self {
         let _camera = Camera::new(
             &mut control_set_builder,
@@ -37,7 +38,7 @@ impl Chart {
             camera: _camera,
             render_targets,
             buffer_generators,
-            passes,
+            steps: passes,
             controls,
         }
     }
@@ -49,7 +50,7 @@ impl Chart {
         for buffer_generator in &self.buffer_generators {
             buffer_generator.generate()?;
         }
-        for pass in &self.passes {
+        for pass in &self.steps {
             pass.render(context, MaterialStepType::Solid, &self.camera)?;
         }
         Ok(())
